requirements_news.txt
feedparser>=6.0.11
requests>=2.31.0
googletrans==4.0.0rc1
google-generativeai>=0.8.3
httpx==0.13.3
python-dotenv>=0.19.0
